separated styled components from react components and refactored styles to be modular across components
