+++ client/src/components/MainHeader.js
@@ -1,17 +1,17 @@
-import styled from "styled-components";
-import { blue3 } from "../variables/colours";
+// import styled from "styled-components";
+// import { blue3 } from "../variables/colours";
 
-const MainHeader = (props) => {
-  const { heading } = props;
+// const MainHeader = (props) => {
+//   const { heading } = props;
 
-  return <Header>{heading}</Header>;
-};
+//   return <Header>{heading}</Header>;
+// };
 
-export default MainHeader;
+// export default MainHeader;
 
-const Header = styled.h2`
-  padding-bottom: 10px;
-  border-bottom: 2px solid ${blue3};
-  margin-bottom: 20px;
-  color: ${blue3};
-`;
+// const Header = styled.h2`
+//   padding-bottom: 10px;
+//   border-bottom: 2px solid ${blue3};
+//   margin-bottom: 20px;
+//   color: ${blue3};
+// `;
