+++ client/src/components/header/Header.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import Logo from "../Logo";
+import Logo from "./Logo";
 import { tabletWidthLrg } from "../../variables/screen";
 import { IconButton } from "../styledComponents/Buttons.styles";
 import { black, blue3 } from "../../variables/colours";
