+++ client/src/components/header/Logo.js
@@ -1,6 +1,6 @@
 import styled from "styled-components";
-import { blue2 } from "../variables/colours";
-import { stdSpace } from "../variables/spacing";
+import { blue2 } from "../../variables/colours";
+import { stdSpace } from "../../variables/spacing";
 
 const Logo = () => {
   return <Name>planIt</Name>;
