+++ client/src/components/styledComponents/ProjectCardWrapper.styles.js
@@ -1,12 +1,12 @@
-import styled from "styled-components";
+// import styled from "styled-components";
 
-const ProjectCardWrapper = styled.div`
-  height: 100px;
-  padding: 10px;
-  text-align: left;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-`;
+// const ProjectCardWrapper = styled.div`
+//   height: 100px;
+//   padding: 10px;
+//   text-align: left;
+//   display: flex;
+//   flex-direction: column;
+//   justify-content: space-between;
+// `;
 
-export default ProjectCardWrapper;
+// export default ProjectCardWrapper;
